defaulting to zero truncates, so do what std::string() does and default to max size_t
svn:revision:52599
svn:branch:trunk
svn:account:brlcad

+++ src/other/step/src/cldai/sdaiString.cc
@@ -13,7 +13,7 @@
 #include <sstream>
 #include "scl_memmgr.h"
 
-SDAI_String::SDAI_String( const char * str, int max ) {
+SDAI_String::SDAI_String( const char * str, size_t max ) {
     content = std::string( str, max );
 }
